Move commander data parsing to Utils
